Realizando crud operadores

+++ demo/src/main/java/com/example/demo/repositorio/OperadorRepository.java
@@ -1,9 +1,9 @@
 package com.example.demo.repositorio;
+
 import com.example.demo.entidades.Operador;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface OperadorRepository extends JpaRepository<Operador, Long> {
+public interface OperadorRepository extends JpaRepository<Operador, Integer> {
 }
